Bug 1408100 - Remove angular-ui-bootstrap from Custom Job Actions dialogs
