+++ ui/entry-index.js
@@ -42,6 +42,7 @@ require('./js/services/log.js');
 require('./js/providers.js');
 require('./js/values.js');
 require('./js/components/auth.js');
+require('./js/components/tcjobactions.js');
 require('./js/directives/treeherder/main.js');
 require('./js/directives/treeherder/clonejobs.js');
 require('./js/directives/treeherder/resultsets.js');
@@ -85,7 +86,6 @@ require('./js/controllers/notification.js');
 require('./js/controllers/filters.js');
 require('./js/controllers/jobs.js');
 require('./js/controllers/bugfiler.js');
-require('./js/controllers/tcjobactions.js');
 require('./plugins/tabs.js');
 require('./plugins/controller.js');
 require('./plugins/pinboard.js');
