+++ ui/index.html
@@ -39,6 +39,7 @@
         </div>
 
         <th-notification-box></th-notification-box>
+        <tc-job-actions></tc-job-actions>
 
         <!-- Clone targets -->
